+++ app/src/main/java/com/maru/inunavi/ui/satisfied/OverviewInfo.java
@@ -97,7 +97,7 @@ public class OverviewInfo {
         return distance;
     }
 
-    public void setDistance(int distance) {
+    public void setDistance(double distance) {
         this.distance = distance;
     }
